# openapi.yaml
openapi: '3.1.0'
info:
  title: 'Mission Control API'
  version: '1.0.0'
  description: |-
    Generated from the Backend’s Fastify route table by `pnpm api:spec` (F5.1). Do not edit
    by hand: `pnpm api:spec:check` fails when this file and the routes disagree, and
    `pnpm api:types:check` fails when apps/frontend/src/lib/api/types.generated.ts does.

    REQUEST shapes are the schemas Ajv actually enforces, so they are exact.

    RESPONSE shapes are the schemas each route declares under `schema.response`. Read the
    guarantee precisely: they are DECLARATIONS, NOT SERIALIZERS. The Backend installs a
    serializer compiler that ignores them and calls JSON.stringify, so a schema here can never
    remove a field the handler produced — the failure mode a Fastify response schema normally
    carries. What keeps them true instead is a conformance check that validates every reply
    against its own schema in the test tiers, in both directions: an undeclared property and a
    missing required property are both failures. See apps/backend/src/http/response-schema.ts.

    Operations still marked `x-mc-response-schema: undeclared` are the ones with no declaration,
    and the marker exists so that gap stays countable. Today they are: GET /health,
    GET /services/health and GET /schedule (owned by modules outside this change),
    GET|PUT /settings/{category} (the GET serves one of five different documents, which needs a
...
    and GET /ws.

...
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /repositories:
    get:
      operationId: getRepositories
      summary: 'GET /repositories'
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Repository'
                  meta:
                    $ref: '#/components/schemas/ListEnvelopeMeta'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    post:
      operationId: postRepositories
      summary: 'POST /repositories'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - localPath
              additionalProperties: false
              properties:
                localPath:
                  type: string
                  minLength: 1
                  maxLength: 4096
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                projectId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                defaultBranch:
                  type: string
                  minLength: 1
                  maxLength: 255
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Repository'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  /repositories/discover:
    post:
      operationId: postRepositoriesDiscover
      summary: 'POST /repositories/discover'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/DiscoveryReport'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/repositories/{id}':
    get:
      operationId: getRepositoriesById
      summary: 'GET /repositories/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Repository'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    patch:
      operationId: patchRepositoriesById
      summary: 'PATCH /repositories/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                projectId:
                  type:
                    - string
                    - 'null'
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                defaultBranch:
                  type: string
                  minLength: 1
                  maxLength: 255
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Repository'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
    delete:
      operationId: deleteRepositoriesById
      summary: 'DELETE /repositories/{id}'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '204':
          description: 'No Content'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
  '/repositories/{id}/status':
    get:
      operationId: getRepositoriesByIdStatus
      summary: 'GET /repositories/{id}/status'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/RepositoryStatus'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
  '/repositories/{id}/sync':
    post:
      operationId: postRepositoriesByIdSync
      summary: 'POST /repositories/{id}/sync'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        '202':
          description: 'Accepted — the work is queued; the outcome is observable on the resource it changes.'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/JobAccepted'
        default:
          $ref: '#/components/responses/ErrorEnvelope'
      x-mc-response-schema: declared
      x-mc-auth:
        mode: authenticated
        scope: full
        allowCookie: true
      x-mc-strictness:
        query: strict
        body: strict
...
components:
  schemas:
...
    DiscoveredRepository:
      type: object
      additionalProperties: false
      required:
        - repository
        - owner
        - repo
      properties:
        repository:
          $ref: '#/components/schemas/Repository'
        owner:
          type: string
        repo:
          type: string
    DiscoveryReport:
      type: object
      additionalProperties: false
      required:
        - scannedAt
        - truncated
        - roots
        - registered
        - skipped
        - counts
        - workflowMode
        - capability
      properties:
        scannedAt:
          type: string
        truncated:
          type: boolean
        roots:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveryRoot'
        registered:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveredRepository'
        skipped:
          type: array
          items:
            $ref: '#/components/schemas/DiscoverySkip'
        counts:
          type: object
          additionalProperties: false
          required:
            - workingTreesFound
            - registered
            - skipped
          properties:
            workingTreesFound:
              type: integer
            registered:
              type: integer
            skipped:
              type: integer
        workflowMode:
          $ref: '#/components/schemas/GithubWorkflowMode'
        capability:
          type: string
          enum:
            - read_only
    DiscoveryRoot:
      type: object
      additionalProperties: false
      required:
        - path
        - status
        - found
        - detail
      properties:
        path:
          type: string
        status:
          $ref: '#/components/schemas/DiscoveryRootStatus'
        found:
          type: integer
        detail:
          type:
            - string
            - 'null'
    DiscoveryRootStatus:
      type: string
      enum:
        - scanned
        - scanned_as_repository
        - path_missing
        - not_a_directory
        - not_absolute
        - unreadable
        - truncated
    DiscoverySkip:
      type: object
      additionalProperties: false
      required:
        - localPath
        - reason
        - detail
        - repositoryId
      properties:
        localPath:
          type: string
        reason:
          $ref: '#/components/schemas/DiscoverySkipReason'
        detail:
          type:
            - string
            - 'null'
        repositoryId:
          type:
            - string
            - 'null'
    DiscoverySkipReason:
      type: string
      enum:
        - already_registered
        - not_a_git_repository
        - no_remote
        - remote_not_github
        - remote_unreadable
        - invalid_path
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
            - details
            - requestId
          properties:
            code:
              $ref: '#/components/schemas/ErrorEnvelopeCode'
            message:
              type: string
            details:
              type:
                - object
                - 'null'
            requestId:
              type: string
              description: 'Also returned as the X-Request-Id header, and in every log line.'
    ErrorEnvelopeCode:
      type: string
      description: |-
        The F5.4 error-code registry (TDS 04 §1.3), verbatim from the Backend’s http/errors.ts.
        Named ErrorEnvelopeCode rather than ErrorCode because the SPA already owns that name for a
        deliberately wider union: lib/api/errors.ts admits codes the client synthesises for failures
        that never reached the server (NETWORK_ERROR, ABORTED). This is the server’s list only.
      enum:
        - CONFLICT
        - DATABASE_SCHEMA_MISMATCH
        - FORBIDDEN
        - INTEGRATION_NOT_CONFIGURED
        - INTERNAL
        - INVALID_CREDENTIALS
        - INVALID_CURSOR
        - INVALID_STATE_TRANSITION
        - NOT_FOUND
        - NO_TURN_IN_FLIGHT
        - OPERATION_NOT_SUPPORTED
        - ORIGIN_NOT_ALLOWED
        - PAYLOAD_TOO_LARGE
        - RATE_LIMITED
        - RUNTIME_UNAVAILABLE
        - SESSION_NOT_RUNNING
        - UNAUTHORIZED
        - VALIDATION_FAILED
...
    GithubWorkflowMode:
      type: string
      enum:
        - manual
        - assisted
...
    JobAccepted:
      type: object
      additionalProperties: false
      required:
        - jobId
      properties:
        jobId:
          type: string
...
    ListEnvelopeMeta:
      type: object
      additionalProperties: false
      required:
        - nextCursor
        - limit
      properties:
        nextCursor:
          type:
            - string
            - 'null'
        limit:
          type: integer
...
    Repository:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - name
        - localPath
        - remoteUrl
        - visibility
        - defaultBranch
        - lastSyncedAt
        - syncStatus
        - lastSyncError
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        projectId:
          type:
            - string
            - 'null'
        name:
          type: string
        localPath:
          type: string
        remoteUrl:
          type:
            - string
            - 'null'
        visibility:
          $ref: '#/components/schemas/RepositoryVisibility'
        defaultBranch:
          type: string
        lastSyncedAt:
          type:
            - string
            - 'null'
        syncStatus:
          $ref: '#/components/schemas/RepositorySyncStatus'
        lastSyncError:
          type:
            - string
            - 'null'
          description: 'Why the last sync failed. Additive to TDS 04 5.1 and served so the Repositories view can explain a `failed` badge without sending the operator to the audit log.'
        createdAt:
          type: string
        updatedAt:
          type: string
    RepositoryStatus:
      type: object
      additionalProperties: false
      required:
        - repositoryId
        - localPath
        - isGitWorkingTree
        - currentBranch
        - detachedHead
        - headSha
        - uncommittedFiles
        - ahead
        - behind
        - unavailableReason
        - detail
        - checkedAt
      properties:
        repositoryId:
          type: string
        localPath:
          type: string
        isGitWorkingTree:
          type: boolean
        currentBranch:
          type:
            - string
            - 'null'
        detachedHead:
          type: boolean
        headSha:
          type:
            - string
            - 'null'
        uncommittedFiles:
          type:
            - integer
            - 'null'
        ahead:
          type:
            - integer
            - 'null'
        behind:
          type:
            - integer
            - 'null'
        unavailableReason:
          anyOf:
            - $ref: '#/components/schemas/WorkingTreeUnavailableReason'
            - type: 'null'
          description: 'null iff the tree was read. Otherwise why not - an answer, not an error. Every value must render as *unverifiable*, never as clean.'
        detail:
          type:
            - string
            - 'null'
        checkedAt:
          type: string
    RepositorySyncStatus:
      type: string
      enum:
        - ok
        - failed
        - never
    RepositoryVisibility:
      type: string
      enum:
        - public
        - private
        - unknown
...
    WorkingTreeUnavailableReason:
      type: string
      enum:
        - path_missing
        - not_a_directory
        - not_a_git_repository
        - git_unavailable
        - timed_out
        - git_failed
  responses:
    ErrorEnvelope:
      description: 'Any non-2xx response (F5.4). The status code and the `error.code` registry entry carry the meaning; see docs/tds/04-api-contracts-and-events.md §1.3.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mc_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'mct_ API token'